requirements.txt
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
astor==0.8.1
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.13.4
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1749229842835/work
cachetools==6.1.0
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1752493286596/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1725560520483/work
chardet==5.2.0
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1746214863626/work
colorlog==6.9.0
contourpy @ file:///home/conda/feedstock_root/build_artifacts/contourpy_1744743067588/work
cssselect==1.3.0
cssutils==2.11.1
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1733332471406/work
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
easyocr==1.7.2
einops==0.8.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1741969488311/work
filterpy==1.4.5
fonttools @ file:///home/conda/feedstock_root/build_artifacts/fonttools_1752722674086/work
frozenlist==1.7.0
fsspec==2025.7.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.45
gmpy2 @ file:///home/conda/feedstock_root/build_artifacts/gmpy2_1745509363766/work
GPUtil==1.4.0
greenlet==3.2.3
h11==0.16.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
hf-xet==1.1.5
hf_transfer==0.1.9
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.3
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
imageio==2.37.0
imagesize==1.4.1
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver @ file:///home/conda/feedstock_root/build_artifacts/kiwisolver_1751493866021/work
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langsmith==0.4.9
lazy_loader==0.4
lxml==6.0.0
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1733219680183/work
marshmallow==3.26.1
matplotlib==3.10.3
more-itertools==10.7.0
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1733302684489/work
multidict==6.6.3
munkres==1.1.4
mypy_extensions==1.1.0
networkx @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_networkx_1731521053/work
ninja==1.11.1.4
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1747544640217/work/dist/numpy-2.2.6-cp310-cp310-linux_x86_64.whl#sha256=d6d964caeef85d00073d27cd62b46883d275b3d8162f723f0fcabbd0b3cc3f9d
nvidia-cublas-cu12==12.4.2.65
nvidia-cuda-cupti-cu12==12.4.99
nvidia-cuda-nvrtc-cu12==12.4.99
nvidia-cuda-runtime-cu12==12.4.99
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.0.44
nvidia-curand-cu12==10.3.5.119
nvidia-cusolver-cu12==11.6.0.99
nvidia-cusparse-cu12==12.3.0.142
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.4.99
openai==1.98.0
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0
opencv-python-headless==4.12.0
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.11.1
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
paddleocr==3.1.0
paddlepaddle==3.1.0
paddlex==3.1.3
pandas==2.3.1
pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1751481981736/work
premailer==3.10.0
prettytable==3.16.0
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pyparsing_1753873557/work
pypdfium2==4.30.0
PySide6==6.9.1
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-bidi==0.6.6
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-dotenv==1.1.1
pytz==2025.2
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1737454647378/work
regex==2025.7.34
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1749498106507/work
requests-toolbelt==1.0.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.1
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy-split_1739790642651/work/dist/scipy-1.15.2-cp310-cp310-linux_x86_64.whl#sha256=9e52bad6c3294d1a5b04a13632118ca2157130603c6c018c2d710162b223b27e
seaborn==0.13.2
sentencepiece==0.2.0
shapely==2.1.1
shiboken6==6.9.1
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.42
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1745946051654/work
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.9.0
tokenizers==0.21.4
torch==2.4.0+cu124
torchaudio==2.4.0
torchvision==0.19.0+cu124
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1748003301700/work
tqdm==4.67.1
transformers==4.54.1
triton==3.0.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1751643513/work
tzdata==2025.2
ujson==5.10.0
ultralytics==8.3.170
ultralytics-thop==2.0.14
unicodedata2 @ file:///home/conda/feedstock_root/build_artifacts/unicodedata2_1736692496989/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1750271362675/work
wcwidth==0.2.13
yarl==1.20.1
zstandard==0.23.0
